Add tests without namespace

+++ tests/run.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
 
 echo flush_all | nc localhost 11211 | grep OK > /dev/null
+if [ $? != 0 ]; then
+	echo "Memcached not ready"
+	exit 1
+fi
 
 rm -rf work
 mkdir work
@@ -21,7 +25,7 @@ cp nginx.conf work
 $NGINX_BIN -p $(pwd)/work -c nginx.conf
 sleep 1
 
-ruby cache_test.rb
+ruby simple_test.rb && ruby ns_test.rb
 res=$?
 
 kill $(cat work/nginx.pid)
